TAD Juego: Ahora si la selección excede los bordes del tablero aparece por el lado opuesto

+++ src/Juego.h
@@ -17,7 +17,7 @@ const int PTOS_PAREJA = 10; 					//Ptos obtenidos al hacer pareja
 const int PTOS_FINAL_TABLERO_VACIO = 50; 		//Ptos de bonificación por vaciar el tablero
 const float TIEMPO_TRANSICION_ESTADOS = 0.25; 	//Retraso entre estados de la animación de voltear fichas
 const float TIEMPO_RETRASO_SELECCION = 0.5; 	//Tiempo que se le deja al jugador ver el anverso de ambas fichas
-const int float TIEMPO_PAUSA_FINAL = 2; 		//Tiempo que se le deja al jugador ver el mensaje final
+const float TIEMPO_PAUSA_FINAL = 2; 		//Tiempo que se le deja al jugador ver el mensaje final
 
 /*
  * PRE:
